Thumbnail refactor WIP: service, types
