+++ src/app/_services/index.ts
@@ -3,3 +3,4 @@ export * from './toast.service'
 export * from './slides.service'
 export * from './collection.service'
 export * from './institution.service'
+export * from './thumbnail.service'
