+++ src/app/shared/datatypes/thumbnail.interface.ts
@@ -47,9 +47,56 @@ export interface Thumbnail {
   }
 }
 
-// export interface SearchThumbnail {
-//   artstorid: string
-//   collections: string[]
-//   collectiontypenameid: string[]
-//   collectiontypes: number[]
+// export interface SearchServiceThumbnail {
+  // id	664db63b-d45e-3ad3-aad2-a3828dc819f2
+  // doi	10.2307/artstor.12240886
+  // debug_fields	{}
+  // agent	Pollock
+  // artdiscoverysite	[]
+  // artstorid	SCHLES_130736960
+  // artadditionalfields	null
+  // clusterid	null
+  // categoryid	1034302695
+  // collectiontypes	[…]
+  // collectiontypenameid	[…]
+  // collections	[…]
+  // compound_media	null
+  // contributinginstitutionid	1000
+  // date	n.d. (circa 1850-1880)
+  // donatinginstitutionids	[]
+  // frequentlygroupedwith	[]
+  // iap	false
+  // media	{"format":null,"thumbnailSizeOnePath":"imgstor/size1/schles/d0001/sch_007133010551_img0068.jpg","width":2048,"sizeInBytes":null,"downloadSize":1024,"type":null,"icc_profile_location":null,"thumbnailSizeZeroPath":"imgstor/size0/schles/d0001/sch_007133010551_img0068.jpg","filename":"SCHLES_130736960","lps":"schles/d0001","iiif":null,"storId":null,"adlObjectType":10,"height":3072}
+  // name	Unidentified portraits of Boit family women.
+  // partofcluster	false
+  // tokens	[…]
+  // type	art
+  // updatedon	2017-12-14T23:12:25Z
+  // workid	null
+  // year	1865
+  // yearbegin	1850
+  // yearend	1880
+  // additional_Fields	{}
+// }
+
+// export interface GroupServiceThumbnail {
+//   objectId	ASPENCEIG_10313206874
+//   compoundmediaCount	0
+//   downloadSize	1024,1024
+//   tombstone	[…]
+//   0	Untitled (spider)
+//   1	Andrew Spence
+//   2	2004
+//   collections	[…]
+//   0	35953
+//   thumbnailImgUrl	/thumb/imgstor/size0/aspence/d0001/spence_untitled_spider_as_8b_srgb.jpg
+//   iap	false
+//   status	available
+//   objectTypeId	10
+//   collectiontypes	[…]
+//   0	1
+//   contributinginstitutionid	1000
+//   clustered	0
+//   largeImgUrl	/thumb/imgstor/size1/aspence/d0001/spence_untitled_spider_as_8b_srgb.jpg
+//   cfObjectId	null
 // }
